python3
Ensure i_max is integer

+++ pycroscopy/analysis/utils/be_sho.py
@@ -123,7 +123,7 @@ def SHOfastGuess(w_vec, resp_vec, qual_factor=200):
         SHO fit parameters arranged as [amplitude, frequency, quality factor, phase]
     """
     amp_vec = abs(resp_vec)
-    i_max = len(resp_vec)/2
+    i_max = int(len(resp_vec)/2)
     return np.array([np.mean(amp_vec) / qual_factor, w_vec[i_max], qual_factor, np.angle(resp_vec[i_max])])
 
 def SHOlowerBound(w_vec):
